added additional layer of abstraction to allow PropertyResolutionFactory to be used outside of classes
