+++ src/model/framework/decorators/Property.ts
@@ -1,11 +1,9 @@
 import { container } from "tsyringe";
 
 import type { propTypes } from "../../Typeings.js";
-import type { Property } from "../engine/IPropertyResolutionEngine.js";
-import { PropertyResolutionFactory } from "../factory/impl/PropertyResolutionFactory.js";
+import { PropertyResolutionManager } from "../manager/PropertyResolutionManager.js";
 
-const factory = container.resolve(PropertyResolutionFactory);
-const engines = factory.engines;
+const manager = container.resolve(PropertyResolutionManager);
 
 /**
  * Get a property from the system. The location where the property is loaded from is agnostic and defined by the registered IPropertyResolutionEngine classes.
@@ -19,21 +17,14 @@ export function Property(prop: keyof propTypes, required = true): PropertyDecora
             configurable: true,
             enumerable: true,
             get: () => {
-                let propValue: Property = null;
-                for (const resolutionEngine of engines) {
-                    const resolvedProp = resolutionEngine.getProperty(prop);
-                    if (resolvedProp !== null) {
-                        propValue = resolvedProp ?? null;
-                        break;
-                    }
-                }
+                const propValue = manager.getProperty(prop);
                 if (required && propValue === null) {
                     throw new Error(`Unable to find prop with key "${prop}"`);
                 }
                 if (!required && propValue === null && original !== null && original !== undefined) {
                     return original;
                 }
-                return propValue ?? null;
+                return propValue;
             },
             set: (newVal) => {
                 original = newVal;
